requirements.txt
flask==2.0.1
torch==1.9.0
torchvision==0.10.0
opencv-python-headless==4.5.3.56
numpy==1.21.2
pandas==1.3.3
requests==2.26.0
ultralytics==8.0.20
Pillow==8.4.0